Changing idnumber field to bigint

+++ database/migrations/2016_01_07_091448_add_columns_to_users_table.php
@@ -14,7 +14,7 @@ class AddColumnsToUsersTable extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->string('lastname');
-	    $table->integer('idnumber');
+	    $table->bigInteger('idnumber');
 	    $table->string('idfilepath')->nullable();
         });
     }
